# internal/delivery/docs/swagger.yaml
definitions:
  entities.User:
    properties:
      id:
        type: integer
      name:
        allOf:
        - $ref: '#/definitions/null.String'
        description: эта библиотека спасает, когда происходит попытка Scan'ировать
          null-value в Goшный тип данных
    type: object
  entities.UserCreate:
    properties:
      name:
        allOf:
        - $ref: '#/definitions/null.String'
        description: эта библиотека спасает, когда происходит попытка Scan'ировать
          null-value в Goшный тип данных
      password:
        type: string
    type: object
  null.String:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /public/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: User create
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created user, returning JWT and Session
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user
      tags:
      - public
  /public/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Mobile user login
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entities.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully loginned user, returning JWT and Session
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login in mobile user
      tags:
      - public
  /public/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: header
        name: Session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authorized, returning JWT and new session_id
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh tokens
      tags:
      - public
  /user/delete:
    get:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: header
        name: Session
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully response
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: JWT is absent or invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: JWT is invalid or expired
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - user
  /user/me:
    get:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: header
        name: Session
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully response with user data
          schema:
            items:
              $ref: '#/definitions/entities.User'
            type: array
        "400":
          description: JWT is absent or invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: JWT is invalid or expired
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user data
      tags:
      - user
swagger: "2.0"